[incubator-kie-issues-1110] Add capability to store that node instance
