+++ jbpm/jbpm-flow/src/main/java/org/jbpm/workflow/instance/node/CompositeNodeInstance.java
@@ -468,6 +468,7 @@ public class CompositeNodeInstance extends StateBasedNodeInstance implements Nod
         this.currentLevel = currentLevel;
     }
 
+    @Override
     public Map<String, Integer> getIterationLevels() {
         return iterationLevels;
     }
