+++ jbpm/process-serialization-protobuf/src/main/java/org/jbpm/flow/serialization/NodeInstanceReader.java
@@ -18,14 +18,23 @@
  */
 package org.jbpm.flow.serialization;
 
-import java.io.OutputStream;
-
 import org.kie.api.runtime.process.NodeInstance;
 
-public interface MarshallerWriterContext extends MarshallerContext {
+import com.google.protobuf.Any;
 
-    OutputStream output();
+public interface NodeInstanceReader extends Comparable<NodeInstanceReader> {
+    Integer DEFAULT_ORDER = 10;
 
-    NodeInstanceWriter findNodeInstanceWriter(NodeInstance nodeInstance);
+    default Integer order() {
+        return DEFAULT_ORDER;
+    }
 
+    @Override
+    default int compareTo(NodeInstanceReader that) {
+        return that.order().compareTo(this.order());
+    }
+
+    boolean accept(Any value);
+
+    NodeInstance read(MarshallerReaderContext context, Any value);
 }
