+++ jbpm/process-serialization-protobuf/src/main/java/org/jbpm/flow/serialization/impl/ProtobufProcessInstanceReader.java
@@ -21,83 +21,42 @@ package org.jbpm.flow.serialization.impl;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.URI;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import org.jbpm.flow.serialization.MarshallerContextName;
 import org.jbpm.flow.serialization.MarshallerReaderContext;
-import org.jbpm.flow.serialization.ProcessInstanceMarshallerException;
+import org.jbpm.flow.serialization.NodeInstanceReader;
 import org.jbpm.flow.serialization.ProcessInstanceMarshallerListener;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.AsyncEventNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.CompositeContextNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.DynamicNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.EventNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.EventSubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.ForEachNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.JoinNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.LambdaSubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.MilestoneNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.RuleSetNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.StateNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.SubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.TimerNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.WorkItemNodeInstanceContent;
 import org.jbpm.flow.serialization.protobuf.KogitoProcessInstanceProtobuf;
 import org.jbpm.flow.serialization.protobuf.KogitoTypesProtobuf;
 import org.jbpm.flow.serialization.protobuf.KogitoTypesProtobuf.SLAContext;
 import org.jbpm.flow.serialization.protobuf.KogitoTypesProtobuf.WorkflowContext;
-import org.jbpm.flow.serialization.protobuf.KogitoWorkItemsProtobuf;
-import org.jbpm.flow.serialization.protobuf.KogitoWorkItemsProtobuf.HumanTaskWorkItemData;
 import org.jbpm.process.core.context.exclusive.ExclusiveGroup;
 import org.jbpm.process.core.context.swimlane.SwimlaneContext;
 import org.jbpm.process.core.context.variable.VariableScope;
+import org.jbpm.process.instance.ContextInstanceContainer;
+import org.jbpm.process.instance.ContextableInstance;
 import org.jbpm.process.instance.context.exclusive.ExclusiveGroupInstance;
 import org.jbpm.process.instance.context.swimlane.SwimlaneContextInstance;
 import org.jbpm.process.instance.context.variable.VariableScopeInstance;
-import org.jbpm.process.instance.impl.humantask.HumanTaskWorkItemImpl;
-import org.jbpm.process.instance.impl.humantask.InternalHumanTaskWorkItem;
-import org.jbpm.process.instance.impl.humantask.Reassignment;
 import org.jbpm.ruleflow.core.WorkflowElementIdentifierFactory;
 import org.jbpm.ruleflow.instance.RuleFlowProcessInstance;
-import org.jbpm.workflow.core.node.AsyncEventNodeInstance;
+import org.jbpm.workflow.instance.NodeInstanceContainer;
 import org.jbpm.workflow.instance.impl.NodeInstanceImpl;
-import org.jbpm.workflow.instance.node.CompositeContextNodeInstance;
-import org.jbpm.workflow.instance.node.DynamicNodeInstance;
-import org.jbpm.workflow.instance.node.EventNodeInstance;
-import org.jbpm.workflow.instance.node.EventSubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.ForEachNodeInstance;
-import org.jbpm.workflow.instance.node.HumanTaskNodeInstance;
-import org.jbpm.workflow.instance.node.JoinInstance;
-import org.jbpm.workflow.instance.node.LambdaSubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.MilestoneNodeInstance;
-import org.jbpm.workflow.instance.node.RuleSetNodeInstance;
-import org.jbpm.workflow.instance.node.StateNodeInstance;
-import org.jbpm.workflow.instance.node.SubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.TimerNodeInstance;
-import org.jbpm.workflow.instance.node.WorkItemNodeInstance;
-import org.kie.api.definition.process.WorkflowElementIdentifier;
 import org.kie.kogito.internal.process.runtime.KogitoNodeInstance;
 import org.kie.kogito.internal.process.runtime.KogitoNodeInstanceContainer;
 import org.kie.kogito.internal.process.runtime.KogitoProcessRuntime;
 import org.kie.kogito.process.impl.AbstractProcess;
-import org.kie.kogito.process.workitem.Attachment;
-import org.kie.kogito.process.workitem.Comment;
-import org.kie.kogito.process.workitems.InternalKogitoWorkItem;
-import org.kie.kogito.process.workitems.impl.KogitoWorkItemImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.protobuf.Any;
-import com.google.protobuf.InvalidProtocolBufferException;
+import com.google.protobuf.Descriptors.FieldDescriptor;
+import com.google.protobuf.GeneratedMessageV3;
 import com.google.protobuf.util.JsonFormat;
 
 import static org.jbpm.flow.serialization.protobuf.ProtobufTypeRegistryFactory.protobufTypeRegistryFactoryInstance;
@@ -114,7 +73,7 @@ public class ProtobufProcessInstanceReader {
         this.ruleFlowProcessInstance = new RuleFlowProcessInstance();
         this.varReader = new ProtobufVariableReader(context);
         this.listeners = context.get(MarshallerContextName.MARSHALLER_INSTANCE_LISTENER);
-        this.listeners = this.listeners != null ? this.listeners : new ProcessInstanceMarshallerListener[0];
+        context.set(MarshallerContextName.MARSHALLER_PROCESS_INSTANCE, ruleFlowProcessInstance);
     }
 
     public RuleFlowProcessInstance read(InputStream input) throws IOException {
@@ -212,25 +171,8 @@ public class ProtobufProcessInstanceReader {
         }
 
         WorkflowContext workflowContext = processInstanceProtobuf.getContext();
+        buildWorkflowContext(processInstance, workflowContext);
 
-        for (KogitoTypesProtobuf.NodeInstance nodeInstanceProtobuf : workflowContext.getNodeInstanceList()) {
-            buildNodeInstance(nodeInstanceProtobuf, processInstance);
-        }
-
-        for (KogitoTypesProtobuf.NodeInstanceGroup group : workflowContext.getExclusiveGroupList()) {
-            Function<String, KogitoNodeInstance> finder = nodeInstanceId -> processInstance.getNodeInstance(nodeInstanceId, true);
-            processInstance.addContextInstance(ExclusiveGroup.EXCLUSIVE_GROUP, buildExclusiveGroupInstance(group, finder));
-        }
-
-        processInstance.addContextInstance(VariableScope.VARIABLE_SCOPE, new VariableScopeInstance());
-        if (workflowContext.getVariableCount() > 0) {
-            VariableScopeInstance variableScopeInstance = (VariableScopeInstance) processInstance.getContextInstance(VariableScope.VARIABLE_SCOPE);
-            varReader.buildVariables(workflowContext.getVariableList()).forEach(v -> variableScopeInstance.internalSetVariable(v.getName(), v.getValue()));
-        }
-
-        if (workflowContext.getIterationLevelsCount() > 0) {
-            processInstance.getIterationLevels().putAll(buildIterationLevels(workflowContext.getIterationLevelsList()));
-        }
         KogitoProcessRuntime runtime = ((AbstractProcess<?>) context.get(MarshallerContextName.MARSHALLER_PROCESS)).getProcessRuntime();
         Arrays.stream(listeners).forEach(e -> e.afterUnmarshallProcess(runtime, processInstance));
         return processInstance;
@@ -258,56 +200,22 @@ public class ProtobufProcessInstanceReader {
 
     protected NodeInstanceImpl buildNodeInstance(KogitoTypesProtobuf.NodeInstance nodeInstance, KogitoNodeInstanceContainer parent) {
         final com.google.protobuf.Any nodeContentProtobuf = nodeInstance.getContent();
-        NodeInstanceImpl result = null;
+
+        NodeInstanceReader reader = context.findNodeInstanceReader(nodeContentProtobuf);
+        if (reader == null) {
+            throw new IllegalArgumentException("Unknown node instance");
+        }
+
         try {
-            if (nodeContentProtobuf.is(RuleSetNodeInstanceContent.class)) {
-                RuleSetNodeInstanceContent content = nodeContentProtobuf.unpack(RuleSetNodeInstanceContent.class);
-                result = buildRuleSetNodeInstance(content);
-            } else if (nodeContentProtobuf.is(ForEachNodeInstanceContent.class)) {
-                ForEachNodeInstanceContent content = nodeContentProtobuf.unpack(ForEachNodeInstanceContent.class);
-                result = buildForEachNodeInstance(content, nodeInstance, parent);
-            } else if (nodeContentProtobuf.is(LambdaSubProcessNodeInstanceContent.class)) {
-                LambdaSubProcessNodeInstanceContent content = nodeContentProtobuf.unpack(LambdaSubProcessNodeInstanceContent.class);
-                result = buildLambdaSubProcessNodeInstance(content);
-            } else if (nodeContentProtobuf.is(SubProcessNodeInstanceContent.class)) {
-                SubProcessNodeInstanceContent content = nodeContentProtobuf.unpack(SubProcessNodeInstanceContent.class);
-                result = buildSubProcessNodeInstance(content);
-            } else if (nodeContentProtobuf.is(StateNodeInstanceContent.class)) {
-                StateNodeInstanceContent content = nodeContentProtobuf.unpack(StateNodeInstanceContent.class);
-                result = buildStateNodeInstance(content);
-            } else if (nodeContentProtobuf.is(JoinNodeInstanceContent.class)) {
-                JoinNodeInstanceContent content = nodeContentProtobuf.unpack(JoinNodeInstanceContent.class);
-                result = buildJoinInstance(content);
-            } else if (nodeContentProtobuf.is(TimerNodeInstanceContent.class)) {
-                TimerNodeInstanceContent content = nodeContentProtobuf.unpack(TimerNodeInstanceContent.class);
-                result = buildTimerNodeInstance(content);
-            } else if (nodeContentProtobuf.is(EventNodeInstanceContent.class)) {
-                result = buildEventNodeInstance();
-            } else if (nodeContentProtobuf.is(MilestoneNodeInstanceContent.class)) {
-                MilestoneNodeInstanceContent content = nodeContentProtobuf.unpack(MilestoneNodeInstanceContent.class);
-                result = buildMilestoneNodeInstance(content);
-            } else if (nodeContentProtobuf.is(DynamicNodeInstanceContent.class)) {
-                DynamicNodeInstanceContent content = nodeContentProtobuf.unpack(DynamicNodeInstanceContent.class);
-                result = buildDynamicNodeInstance(content, nodeInstance, parent);
-            } else if (nodeContentProtobuf.is(EventSubProcessNodeInstanceContent.class)) {
-                EventSubProcessNodeInstanceContent content = nodeContentProtobuf.unpack(EventSubProcessNodeInstanceContent.class);
-                result = buildEventSubProcessNodeInstance(content);
-            } else if (nodeContentProtobuf.is(CompositeContextNodeInstanceContent.class)) {
-                CompositeContextNodeInstanceContent content = nodeContentProtobuf.unpack(CompositeContextNodeInstanceContent.class);
-                result = buildCompositeContextNodeInstance(content, nodeInstance, parent);
-            } else if (nodeContentProtobuf.is(WorkItemNodeInstanceContent.class)) {
-                WorkItemNodeInstanceContent content = nodeContentProtobuf.unpack(WorkItemNodeInstanceContent.class);
-                result = buildWorkItemNodeInstance(content);
-            } else if (nodeContentProtobuf.is(AsyncEventNodeInstanceContent.class)) {
-                AsyncEventNodeInstanceContent content = nodeContentProtobuf.unpack(AsyncEventNodeInstanceContent.class);
-                result = buildAsyncEventNodeInstance(content);
-            }
-
-            if (Objects.isNull(result)) {
-                throw new IllegalArgumentException("Unknown node instance");
-            }
-
+            NodeInstanceImpl result = (NodeInstanceImpl) reader.read(context, nodeContentProtobuf);
             setCommonNodeInstanceData(ruleFlowProcessInstance, parent, nodeInstance, result);
+
+            FieldDescriptor fieldDescriptor = getContextField(nodeInstance);
+            if (fieldDescriptor != null) {
+                GeneratedMessageV3 content = nodeContentProtobuf.unpack(GeneratedMessageV3.class);
+                KogitoTypesProtobuf.WorkflowContext workflowContext = (KogitoTypesProtobuf.WorkflowContext) content.getField(fieldDescriptor);
+                buildWorkflowContext((NodeInstanceContainer & ContextInstanceContainer & ContextableInstance) result, workflowContext);
+            }
 
             SLAContext slaNodeInstanceContext = nodeInstance.getSla();
             result.internalSetSlaCompliance(slaNodeInstanceContext.getSlaCompliance());
@@ -328,277 +236,20 @@ public class ProtobufProcessInstanceReader {
         }
     }
 
-    private NodeInstanceImpl buildAsyncEventNodeInstance(AsyncEventNodeInstanceContent content) {
-        AsyncEventNodeInstance nodeInstance = new AsyncEventNodeInstance();
-        nodeInstance.setJobId(content.getJobId());
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildCompositeContextNodeInstance(CompositeContextNodeInstanceContent content, KogitoTypesProtobuf.NodeInstance protoNodeInstance,
-            KogitoNodeInstanceContainer parentContainer) {
-        CompositeContextNodeInstance nodeInstance = new CompositeContextNodeInstance();
-
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String _timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(_timerId);
+    public FieldDescriptor getContextField(GeneratedMessageV3 meesage) {
+        for (FieldDescriptor field : meesage.getDescriptorForType().getFields()) {
+            if ("context".equals(field.getName()) && field.getContainingType().equals(WorkflowContext.getDescriptor())) {
+                return field;
             }
-            nodeInstance.internalSetTimerInstances(timerInstances);
         }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-
-        setCommonNodeInstanceData(ruleFlowProcessInstance, parentContainer, protoNodeInstance, nodeInstance);
-
-        buildWorkflowContext(nodeInstance, content.getContext());
-        return nodeInstance;
+        return null;
     }
 
-    private NodeInstanceImpl buildEventSubProcessNodeInstance(EventSubProcessNodeInstanceContent content) {
-        EventSubProcessNodeInstance nodeInstance = new EventSubProcessNodeInstance();
-
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String _timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(_timerId);
-            }
-            nodeInstance.internalSetTimerInstances(timerInstances);
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-        buildWorkflowContext(nodeInstance, content.getContext());
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildDynamicNodeInstance(DynamicNodeInstanceContent content, KogitoTypesProtobuf.NodeInstance protoNodeInstance,
-            KogitoNodeInstanceContainer parentContainer) {
-        DynamicNodeInstance nodeInstance = new DynamicNodeInstance();
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String _timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(_timerId);
-            }
-            nodeInstance.internalSetTimerInstances(timerInstances);
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-
-        setCommonNodeInstanceData(ruleFlowProcessInstance, parentContainer, protoNodeInstance, nodeInstance);
-        buildWorkflowContext(nodeInstance, content.getContext());
-
-        return nodeInstance;
-
-    }
-
-    private NodeInstanceImpl buildMilestoneNodeInstance(MilestoneNodeInstanceContent content) {
-        MilestoneNodeInstance nodeInstance = new MilestoneNodeInstance();
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String _timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(_timerId);
-            }
-            nodeInstance.internalSetTimerInstances(timerInstances);
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildEventNodeInstance() {
-        return new EventNodeInstance();
-    }
-
-    private NodeInstanceImpl buildTimerNodeInstance(TimerNodeInstanceContent content) {
-        TimerNodeInstance nodeInstance = new TimerNodeInstance();
-        nodeInstance.internalSetTimerId(content.getTimerId());
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildJoinInstance(JoinNodeInstanceContent content) {
-        JoinInstance nodeInstance = new JoinInstance();
-        if (content.getTriggerCount() > 0) {
-            Map<WorkflowElementIdentifier, Integer> triggers = new HashMap<>();
-            for (JoinNodeInstanceContent.JoinTrigger _join : content.getTriggerList()) {
-                LOGGER.info("unmarshalling join {}", _join.getNodeId());
-                triggers.put(WorkflowElementIdentifierFactory.fromExternalFormat(_join.getNodeId()), _join.getCounter());
-            }
-            nodeInstance.internalSetTriggers(triggers);
-        }
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildStateNodeInstance(StateNodeInstanceContent content) {
-        StateNodeInstance nodeInstance = new StateNodeInstance();
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String _timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(_timerId);
-            }
-            nodeInstance.internalSetTimerInstances(timerInstances);
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildSubProcessNodeInstance(SubProcessNodeInstanceContent content) {
-        SubProcessNodeInstance nodeInstance = new SubProcessNodeInstance();
-        nodeInstance.internalSetProcessInstanceId(content.getProcessInstanceId());
-        if (content.getTimerInstanceIdCount() > 0) {
-            List<String> timerInstances = new ArrayList<>();
-            for (String timerId : content.getTimerInstanceIdList()) {
-                timerInstances.add(timerId);
-            }
-            nodeInstance.internalSetTimerInstances(timerInstances);
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildLambdaSubProcessNodeInstance(LambdaSubProcessNodeInstanceContent content) {
-        LambdaSubProcessNodeInstance nodeInstance = new LambdaSubProcessNodeInstance();
-        nodeInstance.internalSetProcessInstanceId(content.getProcessInstanceId());
-        if (content.getTimerInstanceIdCount() > 0) {
-            nodeInstance.internalSetTimerInstances(new ArrayList<>(content.getTimerInstanceIdList()));
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildForEachNodeInstance(ForEachNodeInstanceContent content, KogitoTypesProtobuf.NodeInstance protoNodeInstance, KogitoNodeInstanceContainer parentContainer) {
-        ForEachNodeInstance nodeInstance = new ForEachNodeInstance();
-        nodeInstance.setExecutedInstances(content.getExecutedInstances());
-        nodeInstance.setTotalInstances(content.getTotalInstances());
-        nodeInstance.setHasAsyncInstances(content.getHasAsyncInstances());
-
-        setCommonNodeInstanceData(ruleFlowProcessInstance, parentContainer, protoNodeInstance, nodeInstance);
-        buildWorkflowContext(nodeInstance, content.getContext());
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildRuleSetNodeInstance(RuleSetNodeInstanceContent content) {
-        RuleSetNodeInstance nodeInstance = new RuleSetNodeInstance();
-        nodeInstance.setRuleFlowGroup(content.getRuleFlowGroup());
-        if (content.getTimerInstanceIdCount() > 0) {
-            nodeInstance.internalSetTimerInstances(new ArrayList<>(content.getTimerInstanceIdList()));
-        }
-        if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-            nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-        }
-
-        return nodeInstance;
-    }
-
-    private NodeInstanceImpl buildWorkItemNodeInstance(WorkItemNodeInstanceContent content) {
-        try {
-            WorkItemNodeInstance nodeInstance = instanceWorkItem(content);
-            if (nodeInstance instanceof HumanTaskNodeInstance) {
-                HumanTaskNodeInstance humanTaskNodeInstance = (HumanTaskNodeInstance) nodeInstance;
-                InternalHumanTaskWorkItem workItem = humanTaskNodeInstance.getWorkItem();
-                Any workItemDataMessage = content.getWorkItemData();
-                if (workItemDataMessage.is(HumanTaskWorkItemData.class)) {
-                    HumanTaskWorkItemData workItemData = workItemDataMessage.unpack(HumanTaskWorkItemData.class);
-                    humanTaskNodeInstance.getNotCompletedDeadlineTimers().putAll(buildDeadlines(workItemData.getCompletedDeadlinesMap()));
-                    humanTaskNodeInstance.getNotCompletedReassigments().putAll(buildReassignments(workItemData.getCompletedReassigmentsMap()));
-                    humanTaskNodeInstance.getNotStartedDeadlineTimers().putAll(buildDeadlines(workItemData.getStartDeadlinesMap()));
-                    humanTaskNodeInstance.getNotStartedReassignments().putAll(buildReassignments(workItemData.getStartReassigmentsMap()));
-
-                    if (workItemData.hasTaskName()) {
-                        workItem.setTaskName(workItemData.getTaskName());
-                    }
-                    if (workItemData.hasTaskDescription()) {
-                        workItem.setTaskDescription(workItemData.getTaskDescription());
-                    }
-                    if (workItemData.hasTaskPriority()) {
-                        workItem.setTaskPriority(workItemData.getTaskPriority());
-                    }
-                    if (workItemData.hasTaskReferenceName()) {
-                        workItem.setReferenceName(workItemData.getTaskReferenceName());
-                    }
-                    if (workItemData.hasActualOwner()) {
-                        workItem.setActualOwner(workItemData.getActualOwner());
-                    }
-                    workItem.getAdminUsers().addAll(workItemData.getAdminUsersList());
-                    workItem.getAdminGroups().addAll(workItemData.getAdminGroupsList());
-                    workItem.getPotentialUsers().addAll(workItemData.getPotUsersList());
-                    workItem.getPotentialGroups().addAll(workItemData.getPotGroupsList());
-                    workItem.getExcludedUsers().addAll(workItemData.getExcludedUsersList());
-                    workItem.getComments().putAll(workItemData.getCommentsList().stream().map(this::buildComment).collect(Collectors.toMap(Comment::getId, Function.identity())));
-                    workItem.getAttachments().putAll(workItemData.getAttachmentsList().stream().map(this::buildAttachment).collect(Collectors.toMap(Attachment::getId, Function.identity())));
-
-                }
-
-            }
-
-            nodeInstance.internalSetWorkItemId(content.getWorkItemId());
-            InternalKogitoWorkItem workItem = (InternalKogitoWorkItem) nodeInstance.getWorkItem();
-            workItem.setId(content.getWorkItemId());
-            workItem.setProcessInstanceId(ruleFlowProcessInstance.getStringId());
-            workItem.setName(content.getName());
-            workItem.setState(content.getState());
-            workItem.setDeploymentId(ruleFlowProcessInstance.getDeploymentId());
-            workItem.setProcessInstance(ruleFlowProcessInstance);
-            workItem.setPhaseId(content.getPhaseId());
-            workItem.setPhaseStatus(content.getPhaseStatus());
-            workItem.setStartDate(new Date(content.getStartDate()));
-            if (content.getCompleteDate() > 0) {
-                workItem.setCompleteDate(new Date(content.getCompleteDate()));
-            }
-
-            if (content.getTimerInstanceIdCount() > 0) {
-                nodeInstance.internalSetTimerInstances(new ArrayList<>(content.getTimerInstanceIdList()));
-            }
-            if (!content.getTimerInstanceReferenceMap().isEmpty()) {
-                nodeInstance.internalSetTimerInstancesReference(new HashMap<>(content.getTimerInstanceReferenceMap()));
-            }
-            nodeInstance.internalSetProcessInstanceId(content.getErrorHandlingProcessInstanceId());
-            varReader.buildVariables(content.getVariableList()).forEach(var -> nodeInstance.getWorkItem().getParameters().put(var.getName(), var.getValue()));
-            varReader.buildVariables(content.getResultList()).forEach(var -> nodeInstance.getWorkItem().getResults().put(var.getName(), var.getValue()));
-            return nodeInstance;
-        } catch (InvalidProtocolBufferException ex) {
-            throw new ProcessInstanceMarshallerException("cannot unpack node instance", ex);
-        }
-    }
-
-    private WorkItemNodeInstance instanceWorkItem(WorkItemNodeInstanceContent content) {
-        if (content.hasWorkItemData()) {
-            Any workItemDataMessage = content.getWorkItemData();
-            if (workItemDataMessage.is(HumanTaskWorkItemData.class)) {
-                HumanTaskNodeInstance nodeInstance = new HumanTaskNodeInstance();
-                HumanTaskWorkItemImpl workItem = new HumanTaskWorkItemImpl();
-                nodeInstance.internalSetWorkItem(workItem);
-                return nodeInstance;
-            } else {
-                throw new ProcessInstanceMarshallerException("Don't know which type of work item is");
-            }
-        } else {
-            WorkItemNodeInstance nodeInstance = new WorkItemNodeInstance();
-            KogitoWorkItemImpl workItem = new KogitoWorkItemImpl();
-            workItem.setId(UUID.randomUUID().toString());
-            nodeInstance.internalSetWorkItem(workItem);
-            return nodeInstance;
-        }
-    }
-
-    private void buildWorkflowContext(CompositeContextNodeInstance container, WorkflowContext workflowContext) {
+    private <T extends NodeInstanceContainer & ContextInstanceContainer & ContextableInstance> void buildWorkflowContext(T container, WorkflowContext workflowContext) {
         if (workflowContext.getNodeInstanceCount() > 0) {
             for (KogitoTypesProtobuf.NodeInstance nodeInstanceProtobuf : workflowContext.getNodeInstanceList()) {
                 buildNodeInstance(nodeInstanceProtobuf, container);
             }
-        }
-        for (KogitoTypesProtobuf.NodeInstanceGroup group : workflowContext.getExclusiveGroupList()) {
-            Function<String, KogitoNodeInstance> finder = nodeInstanceId -> container.getNodeInstance(nodeInstanceId, true);
-            container.addContextInstance(ExclusiveGroup.EXCLUSIVE_GROUP, buildExclusiveGroupInstance(group, finder));
         }
 
         container.addContextInstance(VariableScope.VARIABLE_SCOPE, new VariableScopeInstance());
@@ -608,6 +259,10 @@ public class ProtobufProcessInstanceReader {
         }
         if (workflowContext.getIterationLevelsCount() > 0) {
             container.getIterationLevels().putAll(buildIterationLevels(workflowContext.getIterationLevelsList()));
+        }
+        for (KogitoTypesProtobuf.NodeInstanceGroup group : workflowContext.getExclusiveGroupList()) {
+            Function<String, KogitoNodeInstance> finder = nodeInstanceId -> container.getNodeInstance(nodeInstanceId, true);
+            container.addContextInstance(ExclusiveGroup.EXCLUSIVE_GROUP, buildExclusiveGroupInstance(group, finder));
         }
     }
 
@@ -629,39 +284,4 @@ public class ProtobufProcessInstanceReader {
         return iterationLevel.stream().collect(Collectors.toMap(mapKey, mapValue));
     }
 
-    private Comment buildComment(KogitoWorkItemsProtobuf.Comment comment) {
-        Comment result = new Comment(comment.getId(), comment.getUpdatedBy());
-        result.setContent(comment.getContent());
-        result.setUpdatedAt(new Date(comment.getUpdatedAt()));
-        return result;
-    }
-
-    private Attachment buildAttachment(KogitoWorkItemsProtobuf.Attachment attachment) {
-        Attachment result = new Attachment(attachment.getId(), attachment.getUpdatedBy());
-        result.setContent(URI.create(attachment.getContent()));
-        result.setUpdatedAt(new Date(attachment.getUpdatedAt()));
-        result.setName(attachment.getName());
-        return result;
-    }
-
-    private Map<String, Map<String, Object>> buildDeadlines(Map<String, KogitoWorkItemsProtobuf.Deadline> deadlinesProtobuf) {
-        Map<String, Map<String, Object>> deadlines = new HashMap<>();
-        for (Map.Entry<String, KogitoWorkItemsProtobuf.Deadline> entry : deadlinesProtobuf.entrySet()) {
-            Map<String, Object> notification = new HashMap<>();
-            for (Map.Entry<String, String> pair : entry.getValue().getContentMap().entrySet()) {
-                notification.put(pair.getKey(), pair.getValue());
-            }
-            deadlines.put(entry.getKey(), notification);
-        }
-        return deadlines;
-    }
-
-    private Map<String, Reassignment> buildReassignments(Map<String, KogitoWorkItemsProtobuf.Reassignment> reassignmentsProtobuf) {
-        Map<String, Reassignment> reassignments = new HashMap<>();
-        for (Map.Entry<String, KogitoWorkItemsProtobuf.Reassignment> entry : reassignmentsProtobuf.entrySet()) {
-            reassignments.put(entry.getKey(), new Reassignment(entry.getValue().getUsersList().stream().collect(Collectors
-                    .toSet()), entry.getValue().getGroupsList().stream().collect(Collectors.toSet())));
-        }
-        return reassignments;
-    }
 }
