+++ jbpm/process-serialization-protobuf/src/main/java/org/jbpm/flow/serialization/impl/ProtobufProcessInstanceWriter.java
@@ -26,72 +26,39 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.stream.Collectors;
 
 import org.jbpm.flow.serialization.MarshallerContextName;
 import org.jbpm.flow.serialization.MarshallerWriterContext;
+import org.jbpm.flow.serialization.NodeInstanceWriter;
 import org.jbpm.flow.serialization.ProcessInstanceMarshallerListener;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.AsyncEventNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.CompositeContextNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.DynamicNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.EventNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.EventSubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.ForEachNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.JoinNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.JoinNodeInstanceContent.JoinTrigger;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.LambdaSubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.MilestoneNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.RuleSetNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.StateNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.SubProcessNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.TimerNodeInstanceContent;
-import org.jbpm.flow.serialization.protobuf.KogitoNodeInstanceContentsProtobuf.WorkItemNodeInstanceContent;
 import org.jbpm.flow.serialization.protobuf.KogitoProcessInstanceProtobuf;
 import org.jbpm.flow.serialization.protobuf.KogitoTypesProtobuf;
 import org.jbpm.flow.serialization.protobuf.KogitoTypesProtobuf.WorkflowContext;
-import org.jbpm.flow.serialization.protobuf.KogitoWorkItemsProtobuf;
-import org.jbpm.flow.serialization.protobuf.KogitoWorkItemsProtobuf.HumanTaskWorkItemData;
 import org.jbpm.process.core.context.exclusive.ExclusiveGroup;
 import org.jbpm.process.core.context.swimlane.SwimlaneContext;
 import org.jbpm.process.core.context.variable.VariableScope;
 import org.jbpm.process.instance.ContextInstance;
+import org.jbpm.process.instance.ContextInstanceContainer;
+import org.jbpm.process.instance.ContextableInstance;
 import org.jbpm.process.instance.context.exclusive.ExclusiveGroupInstance;
 import org.jbpm.process.instance.context.swimlane.SwimlaneContextInstance;
 import org.jbpm.process.instance.context.variable.VariableScopeInstance;
-import org.jbpm.process.instance.impl.humantask.Reassignment;
-import org.jbpm.workflow.core.node.AsyncEventNodeInstance;
+import org.jbpm.ruleflow.instance.RuleFlowProcessInstance;
+import org.jbpm.workflow.instance.NodeInstanceContainer;
 import org.jbpm.workflow.instance.impl.WorkflowProcessInstanceImpl;
-import org.jbpm.workflow.instance.node.CompositeContextNodeInstance;
-import org.jbpm.workflow.instance.node.DynamicNodeInstance;
-import org.jbpm.workflow.instance.node.EventNodeInstance;
-import org.jbpm.workflow.instance.node.EventSubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.ForEachNodeInstance;
-import org.jbpm.workflow.instance.node.HumanTaskNodeInstance;
-import org.jbpm.workflow.instance.node.JoinInstance;
-import org.jbpm.workflow.instance.node.LambdaSubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.MilestoneNodeInstance;
-import org.jbpm.workflow.instance.node.RuleSetNodeInstance;
-import org.jbpm.workflow.instance.node.StateNodeInstance;
-import org.jbpm.workflow.instance.node.SubProcessNodeInstance;
-import org.jbpm.workflow.instance.node.TimerNodeInstance;
-import org.jbpm.workflow.instance.node.WorkItemNodeInstance;
-import org.kie.api.definition.process.WorkflowElementIdentifier;
 import org.kie.api.runtime.process.NodeInstance;
 import org.kie.kogito.internal.process.runtime.KogitoNodeInstance;
 import org.kie.kogito.internal.process.runtime.KogitoProcessRuntime;
-import org.kie.kogito.internal.process.runtime.KogitoWorkItem;
 import org.kie.kogito.process.impl.AbstractProcess;
-import org.kie.kogito.process.workitem.Attachment;
-import org.kie.kogito.process.workitem.Comment;
-import org.kie.kogito.process.workitem.HumanTaskWorkItem;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.protobuf.Any;
+import com.google.protobuf.Descriptors.FieldDescriptor;
+import com.google.protobuf.GeneratedMessageV3;
 import com.google.protobuf.util.JsonFormat;
 
 import static org.jbpm.flow.serialization.MarshallerContextName.MARSHALLER_FORMAT;
@@ -108,10 +75,11 @@ public class ProtobufProcessInstanceWriter {
         this.context = context;
         this.varWriter = new ProtobufVariableWriter(context);
         this.listeners = context.get(MarshallerContextName.MARSHALLER_INSTANCE_LISTENER);
-        this.listeners = this.listeners != null ? this.listeners : new ProcessInstanceMarshallerListener[0];
     }
 
     public void writeProcessInstance(WorkflowProcessInstanceImpl workFlow, OutputStream os) throws IOException {
+        context.set(MarshallerContextName.MARSHALLER_PROCESS_INSTANCE, (RuleFlowProcessInstance) workFlow);
+        LOGGER.debug("writing process instance {}", workFlow.getId());
         KogitoProcessRuntime runtime = ((AbstractProcess<?>) context.get(MarshallerContextName.MARSHALLER_PROCESS)).getProcessRuntime();
         Arrays.stream(listeners).forEach(e -> e.beforeMarshallProcess(runtime, workFlow));
 
@@ -166,12 +134,7 @@ public class ProtobufProcessInstanceWriter {
 
         instance.addAllSwimlaneContext(buildSwimlaneContexts((SwimlaneContextInstance) workFlow.getContextInstance(SwimlaneContext.SWIMLANE_SCOPE)));
 
-        List<NodeInstance> nodeInstances = new ArrayList<>(workFlow.getNodeInstances());
-        List<ContextInstance> exclusiveGroupInstances = workFlow.getContextInstances(ExclusiveGroup.EXCLUSIVE_GROUP);
-        VariableScopeInstance variableScopeInstance = (VariableScopeInstance) workFlow.getContextInstance(VariableScope.VARIABLE_SCOPE);
-        List<Map.Entry<String, Object>> variables = new ArrayList<>(variableScopeInstance.getVariables().entrySet());
-        List<Map.Entry<String, Integer>> iterationlevels = new ArrayList<>(workFlow.getIterationLevels().entrySet());
-        instance.setContext(buildWorkflowContext(nodeInstances, exclusiveGroupInstances, variables, iterationlevels));
+        instance.setContext(buildWorkflowContext(workFlow));
 
         KogitoProcessInstanceProtobuf.ProcessInstance piProtobuf = instance.build();
 
@@ -251,262 +214,43 @@ public class ProtobufProcessInstanceWriter {
             nodeInstancesProtobuf.add(node.build());
         }
         return nodeInstancesProtobuf;
+
     }
 
     private Any buildNodeInstanceContent(NodeInstance nodeInstance) {
         KogitoProcessRuntime runtime = ((AbstractProcess<?>) context.get(MarshallerContextName.MARSHALLER_PROCESS)).getProcessRuntime();
         Arrays.stream(listeners).forEach(e -> e.beforeMarshallNode(runtime, (KogitoNodeInstance) nodeInstance));
 
-        if (nodeInstance instanceof RuleSetNodeInstance) {
-            return buildRuleSetNodeInstance((RuleSetNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof ForEachNodeInstance) {
-            return buildForEachNodeInstance((ForEachNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof LambdaSubProcessNodeInstance) {
-            return buildLambdaSubProcessNodeInstance((LambdaSubProcessNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof SubProcessNodeInstance) {
-            return buildSubProcessNodeInstance((SubProcessNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof StateNodeInstance) {
-            return buildStateNodeInstance((StateNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof JoinInstance) {
-            return buildJoinInstance((JoinInstance) nodeInstance);
-        } else if (nodeInstance instanceof TimerNodeInstance) {
-            return buildTimerNodeInstance((TimerNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof AsyncEventNodeInstance) {
-            return buildAsyncEventNodeInstance((AsyncEventNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof EventNodeInstance) {
-            return buildEventNodeInstance();
-        } else if (nodeInstance instanceof MilestoneNodeInstance) {
-            return buildMilestoneNodeInstance((MilestoneNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof DynamicNodeInstance) {
-            return buildDynamicNodeInstance((DynamicNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof EventSubProcessNodeInstance) {
-            return buildEventSubProcessNodeInstance((EventSubProcessNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof CompositeContextNodeInstance) {
-            return buildCompositeContextNodeInstance((CompositeContextNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof HumanTaskNodeInstance) {
-            return buildHumanTaskNodeInstance((HumanTaskNodeInstance) nodeInstance);
-        } else if (nodeInstance instanceof WorkItemNodeInstance) {
-            return buildWorkItemNodeInstance((WorkItemNodeInstance) nodeInstance);
-        } else {
+        NodeInstanceWriter writer = context.findNodeInstanceWriter(nodeInstance);
+        if (writer == null) {
             throw new IllegalArgumentException("Unknown node instance type: " + nodeInstance);
         }
-    }
+        GeneratedMessageV3.Builder<?> builder = writer.write(context, nodeInstance);
 
-    private Any buildAsyncEventNodeInstance(AsyncEventNodeInstance nodeInstance) {
-        AsyncEventNodeInstanceContent.Builder builder = AsyncEventNodeInstanceContent.newBuilder();
-        if (nodeInstance.getJobId() != null) {
-            builder.setJobId(nodeInstance.getJobId());
-        }
-        return Any.pack(builder.build());
-    }
-
-    private Any buildRuleSetNodeInstance(RuleSetNodeInstance nodeInstance) {
-        RuleSetNodeInstanceContent.Builder ruleSet = RuleSetNodeInstanceContent.newBuilder();
-        ruleSet.setRuleFlowGroup(nodeInstance.getRuleFlowGroup());
-        ruleSet.addAllTimerInstanceId(nodeInstance.getTimerInstances());
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            ruleSet.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        return Any.pack(ruleSet.build());
-    }
-
-    private Any buildForEachNodeInstance(ForEachNodeInstance nodeInstance) {
-        ForEachNodeInstanceContent.Builder foreachBuilder = ForEachNodeInstanceContent.newBuilder();
-
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            foreachBuilder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            foreachBuilder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        List<NodeInstance> nodeInstances = nodeInstance.getNodeInstances().stream().filter(CompositeContextNodeInstance.class::isInstance).collect(Collectors.toList());
-
-        List<ContextInstance> exclusiveGroupInstances = nodeInstance.getContextInstances(ExclusiveGroup.EXCLUSIVE_GROUP);
-        VariableScopeInstance variableScopeInstance = (VariableScopeInstance) nodeInstance.getContextInstance(VariableScope.VARIABLE_SCOPE);
-        List<Map.Entry<String, Object>> variables = (variableScopeInstance != null) ? new ArrayList<>(variableScopeInstance.getVariables().entrySet()) : Collections.emptyList();
-        List<Map.Entry<String, Integer>> iterationlevels = new ArrayList<>(nodeInstance.getIterationLevels().entrySet());
-        foreachBuilder.setContext(buildWorkflowContext(nodeInstances, exclusiveGroupInstances, variables, iterationlevels));
-
-        foreachBuilder
-                .setTotalInstances(nodeInstance.getTotalInstances())
-                .setExecutedInstances(nodeInstance.getExecutedInstances())
-                .setHasAsyncInstances(nodeInstance.getHasAsyncInstances());
-        return Any.pack(foreachBuilder.build());
-    }
-
-    private Any buildLambdaSubProcessNodeInstance(LambdaSubProcessNodeInstance nodeInstance) {
-        LambdaSubProcessNodeInstanceContent.Builder builder = LambdaSubProcessNodeInstanceContent.newBuilder();
-        builder.setProcessInstanceId(nodeInstance.getProcessInstanceId());
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
+        FieldDescriptor fieldContext = getContextField(builder);
+        if (fieldContext != null) {
+            builder.setField(fieldContext, buildWorkflowContext((NodeInstanceContainer & ContextInstanceContainer & ContextableInstance) nodeInstance));
         }
 
         return Any.pack(builder.build());
     }
 
-    private Any buildSubProcessNodeInstance(SubProcessNodeInstance nodeInstance) {
-        SubProcessNodeInstanceContent.Builder builder = SubProcessNodeInstanceContent.newBuilder();
-        builder.setProcessInstanceId(nodeInstance.getProcessInstanceId());
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
+    public FieldDescriptor getContextField(GeneratedMessageV3.Builder<?> builder) {
+        for (FieldDescriptor field : builder.getDescriptorForType().getFields()) {
+            if ("context".equals(field.getName())) {
+                return field;
+            }
         }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        return Any.pack(builder.build());
+        return null;
     }
 
-    private Any buildStateNodeInstance(StateNodeInstance nodeInstance) {
-        StateNodeInstanceContent.Builder builder = StateNodeInstanceContent.newBuilder();
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        return Any.pack(builder.build());
-    }
-
-    private Any buildJoinInstance(JoinInstance nodeInstance) {
-        JoinNodeInstanceContent.Builder joinBuilder = JoinNodeInstanceContent.newBuilder();
-        Map<WorkflowElementIdentifier, Integer> triggers = nodeInstance.getTriggers();
-        List<WorkflowElementIdentifier> keys = new ArrayList<>(triggers.keySet());
-        Collections.sort(keys);
-
-        for (WorkflowElementIdentifier key : keys) {
-            LOGGER.info("marshalling join {}", key.toExternalFormat());
-            joinBuilder.addTrigger(JoinTrigger.newBuilder()
-                    .setNodeId(key.toExternalFormat())
-                    .setCounter(triggers.get(key))
-                    .build());
-        }
-
-        return Any.pack(joinBuilder.build());
-    }
-
-    private Any buildTimerNodeInstance(TimerNodeInstance nodeInstance) {
-        return Any.pack(TimerNodeInstanceContent.newBuilder().setTimerId(nodeInstance.getTimerId()).build());
-    }
-
-    private Any buildEventNodeInstance() {
-        return Any.pack(EventNodeInstanceContent.newBuilder().build());
-    }
-
-    private Any buildMilestoneNodeInstance(MilestoneNodeInstance nodeInstance) {
-        MilestoneNodeInstanceContent.Builder builder = MilestoneNodeInstanceContent.newBuilder();
-
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        return Any.pack(builder.build());
-    }
-
-    private Any buildDynamicNodeInstance(DynamicNodeInstance nodeInstance) {
-
-        DynamicNodeInstanceContent.Builder builder = DynamicNodeInstanceContent.newBuilder();
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-
-        builder.setContext(buildWorkflowContext(nodeInstance));
-
-        return Any.pack(builder.build());
-    }
-
-    private Any buildEventSubProcessNodeInstance(EventSubProcessNodeInstance nodeInstance) {
-
-        EventSubProcessNodeInstanceContent.Builder builder = EventSubProcessNodeInstanceContent.newBuilder();
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-
-        builder.setContext(buildWorkflowContext(nodeInstance));
-
-        return Any.pack(builder.build());
-    }
-
-    private Any buildCompositeContextNodeInstance(CompositeContextNodeInstance nodeInstance) {
-
-        CompositeContextNodeInstanceContent.Builder builder = CompositeContextNodeInstanceContent.newBuilder();
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-
-        builder.setContext(buildWorkflowContext(nodeInstance));
-
-        return Any.pack(builder.build());
-    }
-
-    private WorkflowContext buildWorkflowContext(CompositeContextNodeInstance nodeInstance) {
+    private <T extends NodeInstanceContainer & ContextInstanceContainer & ContextableInstance> WorkflowContext buildWorkflowContext(T nodeInstance) {
         List<NodeInstance> nodeInstances = new ArrayList<>(nodeInstance.getNodeInstances());
         List<ContextInstance> exclusiveGroupInstances = nodeInstance.getContextInstances(ExclusiveGroup.EXCLUSIVE_GROUP);
         VariableScopeInstance variableScopeInstance = (VariableScopeInstance) nodeInstance.getContextInstance(VariableScope.VARIABLE_SCOPE);
         List<Map.Entry<String, Object>> variables = (variableScopeInstance != null) ? new ArrayList<>(variableScopeInstance.getVariables().entrySet()) : Collections.emptyList();
         List<Map.Entry<String, Integer>> iterationlevels = new ArrayList<>(nodeInstance.getIterationLevels().entrySet());
         return buildWorkflowContext(nodeInstances, exclusiveGroupInstances, variables, iterationlevels);
-    }
-
-    private Any buildWorkItemNodeInstance(WorkItemNodeInstance nodeInstance) {
-        return Any.pack(buildWorkItemNodeInstanceBuilder(nodeInstance).build());
-    }
-
-    private WorkItemNodeInstanceContent.Builder buildWorkItemNodeInstanceBuilder(WorkItemNodeInstance nodeInstance) {
-        WorkItemNodeInstanceContent.Builder builder = WorkItemNodeInstanceContent.newBuilder();
-
-        List<String> timerInstances = nodeInstance.getTimerInstances();
-        if (timerInstances != null) {
-            builder.addAllTimerInstanceId(timerInstances);
-        }
-        if (nodeInstance.getTimerInstancesReference() != null) {
-            builder.putAllTimerInstanceReference(nodeInstance.getTimerInstancesReference());
-        }
-        if (nodeInstance.getExceptionHandlingProcessInstanceId() != null) {
-            builder.setErrorHandlingProcessInstanceId(nodeInstance.getExceptionHandlingProcessInstanceId());
-        }
-        KogitoWorkItem workItem = nodeInstance.getWorkItem();
-
-        builder.setWorkItemId(nodeInstance.getWorkItemId())
-                .setName(workItem.getName())
-                .setState(workItem.getState())
-                .setPhaseId(workItem.getPhaseId())
-                .setPhaseStatus(workItem.getPhaseStatus())
-                .setStartDate(workItem.getStartDate().getTime())
-                .addAllVariable(varWriter.buildVariables(new ArrayList<>(workItem.getParameters().entrySet())))
-                .addAllResult(varWriter.buildVariables(new ArrayList<>(workItem.getResults().entrySet())));
-
-        if (workItem.getCompleteDate() != null) {
-            builder.setCompleteDate(workItem.getCompleteDate().getTime());
-        }
-        return builder;
-    }
-
-    private Any buildHumanTaskNodeInstance(HumanTaskNodeInstance nodeInstance) {
-        WorkItemNodeInstanceContent.Builder builder = buildWorkItemNodeInstanceBuilder(nodeInstance);
-        builder.setWorkItemData(Any.pack(buildHumanTaskWorkItemData(nodeInstance, (HumanTaskWorkItem) nodeInstance.getWorkItem())));
-        return Any.pack(builder.build());
     }
 
     private List<KogitoTypesProtobuf.NodeInstanceGroup> buildGroups(List<ContextInstance> exclusiveGroupInstances) {
@@ -543,123 +287,6 @@ public class ProtobufProcessInstanceWriter {
             }
         }
         return levelsProtobuf;
-    }
-
-    private HumanTaskWorkItemData buildHumanTaskWorkItemData(HumanTaskNodeInstance nodeInstance, HumanTaskWorkItem workItem) {
-        HumanTaskWorkItemData.Builder builder = HumanTaskWorkItemData.newBuilder();
-
-        if (workItem.getTaskPriority() != null) {
-            builder.setTaskPriority(workItem.getTaskPriority());
-        }
-
-        if (workItem.getReferenceName() != null) {
-            builder.setTaskReferenceName(workItem.getReferenceName());
-        }
-        if (workItem.getTaskDescription() != null) {
-            builder.setTaskDescription(workItem.getTaskDescription());
-        }
-
-        if (workItem.getActualOwner() != null) {
-            builder.setActualOwner(workItem.getActualOwner());
-        }
-
-        if (workItem.getTaskName() != null) {
-            builder.setTaskName(workItem.getTaskName());
-        }
-
-        if (workItem.getPotentialUsers() != null) {
-            builder.addAllPotUsers(workItem.getPotentialUsers());
-        }
-
-        if (workItem.getPotentialGroups() != null) {
-            builder.addAllPotGroups(workItem.getPotentialGroups());
-        }
-
-        if (workItem.getExcludedUsers() != null) {
-            builder.addAllExcludedUsers(workItem.getExcludedUsers());
-        }
-
-        if (workItem.getAdminUsers() != null) {
-            builder.addAllAdminUsers(workItem.getAdminUsers());
-        }
-
-        if (workItem.getAdminGroups() != null) {
-            builder.addAllAdminGroups(workItem.getAdminGroups());
-        }
-
-        if (workItem.getComments() != null) {
-            builder.addAllComments(buildComments(workItem.getComments().values()));
-        }
-
-        if (workItem.getAttachments() != null) {
-            builder.addAllAttachments(buildAttachments(workItem.getAttachments().values()));
-        }
-
-        if (nodeInstance.getNotStartedDeadlineTimers() != null) {
-            builder.putAllStartDeadlines(buildDeadlines(nodeInstance.getNotStartedDeadlineTimers()));
-        }
-
-        if (nodeInstance.getNotStartedReassignments() != null) {
-            builder.putAllStartReassigments(buildReassignments(nodeInstance.getNotStartedReassignments()));
-        }
-
-        if (nodeInstance.getNotCompletedDeadlineTimers() != null) {
-            builder.putAllCompletedDeadlines(buildDeadlines(nodeInstance.getNotCompletedDeadlineTimers()));
-        }
-
-        if (nodeInstance.getNotCompletedReassigments() != null) {
-            builder.putAllCompletedReassigments(buildReassignments(nodeInstance.getNotCompletedReassigments()));
-        }
-
-        return builder.build();
-    }
-
-    private List<KogitoWorkItemsProtobuf.Comment> buildComments(Iterable<Comment> comments) {
-        List<KogitoWorkItemsProtobuf.Comment> commentsProtobuf = new ArrayList<>();
-        for (Comment comment : comments) {
-            KogitoWorkItemsProtobuf.Comment workItemComment = KogitoWorkItemsProtobuf.Comment.newBuilder()
-                    .setId(comment.getId().toString())
-                    .setContent(comment.getContent())
-                    .setUpdatedBy(comment.getUpdatedBy())
-                    .setUpdatedAt(comment.getUpdatedAt().getTime())
-                    .build();
-            commentsProtobuf.add(workItemComment);
-        }
-        return commentsProtobuf;
-    }
-
-    private List<KogitoWorkItemsProtobuf.Attachment> buildAttachments(Iterable<Attachment> attachments) {
-        List<KogitoWorkItemsProtobuf.Attachment> attachmentProtobuf = new ArrayList<>();
-        for (Attachment attachment : attachments) {
-            KogitoWorkItemsProtobuf.Attachment workItemAttachment = KogitoWorkItemsProtobuf.Attachment.newBuilder()
-                    .setId(attachment.getId().toString()).setContent(attachment.getContent().toString())
-                    .setUpdatedBy(attachment.getUpdatedBy()).setUpdatedAt(attachment.getUpdatedAt().getTime())
-                    .setName(attachment.getName())
-                    .build();
-            attachmentProtobuf.add(workItemAttachment);
-        }
-        return attachmentProtobuf;
-    }
-
-    private Map<String, KogitoWorkItemsProtobuf.Deadline> buildDeadlines(Map<String, Map<String, Object>> deadlines) {
-        Map<String, KogitoWorkItemsProtobuf.Deadline> deadlinesProtobuf = new HashMap<>();
-        for (Map.Entry<String, Map<String, Object>> entry : deadlines.entrySet()) {
-            KogitoWorkItemsProtobuf.Deadline.Builder builder = KogitoWorkItemsProtobuf.Deadline.newBuilder();
-            entry.getValue().forEach((k, v) -> builder.putContent(k, v.toString()));
-            deadlinesProtobuf.put(entry.getKey(), builder.build());
-        }
-        return deadlinesProtobuf;
-    }
-
-    private Map<String, KogitoWorkItemsProtobuf.Reassignment> buildReassignments(Map<String, Reassignment> reassignments) {
-        Map<String, KogitoWorkItemsProtobuf.Reassignment> reassignmentsProtobuf = new HashMap<>();
-        for (Map.Entry<String, Reassignment> entry : reassignments.entrySet()) {
-            KogitoWorkItemsProtobuf.Reassignment.Builder builder = KogitoWorkItemsProtobuf.Reassignment.newBuilder();
-            builder.addAllGroups(entry.getValue().getPotentialGroups());
-            builder.addAllUsers(entry.getValue().getPotentialUsers());
-            reassignmentsProtobuf.put(entry.getKey(), builder.build());
-        }
-        return reassignmentsProtobuf;
     }
 
 }
